+++ packages/system/src/elevation/types.ts
@@ -1,0 +1,9 @@
+import { elevations } from '@gympass/yoga-tokens';
+
+type ElevationsValues = typeof elevations | string;
+
+export type Elevations = {
+  boxShadow?: ElevationsValues;
+  bs?: ElevationsValues;
+  elevation?: ElevationsValues;
+};