+++ packages/system/src/font-weight/index.ts
@@ -1,0 +1,2 @@
+export * from './fontWeight';
+export type { FontWeights } from './types';