+++ packages/tokens/src/global/index.ts
@@ -1,5 +1,5 @@
 import borders from './borders';
-import breakpoints from './breakpoints';
+import breakpoints, { BREAKPOINTS_KEYS } from './breakpoints';
 import colors from './colors';
 import fonts from './fonts';
 import fontSizes from './font-sizes';
@@ -13,6 +13,7 @@ import transition from './transition';
 const tokens = {
   borders,
   breakpoints,
+  BREAKPOINTS_KEYS,
   colors,
   fonts,
   fontSizes,
@@ -28,6 +29,7 @@ export {
   tokens as default,
   borders,
   breakpoints,
+  BREAKPOINTS_KEYS,
   colors,
   fonts,
   fontSizes,