+++ packages/tokens/src/index.ts
@@ -1,4 +1,6 @@
 import tokens from './global';
 import znCn from './locale/zn-cn';
+import { BreakpointsKey, BreakpointsType } from './global/breakpoints';
 
-export { tokens as default, znCn };
+export { tokens as default, znCn, BreakpointsKey, BreakpointsType };
+export * from './global';