+++ packages/yoga/src/Result/index.native.js
@@ -2,9 +2,11 @@ import Result from './native';
 import ResultDetails from './native/Details';
 import ResultButton from './native/ResultButton';
 import ResultTags from './native/Tags';
+import Badge from './native/Badge';
 
 Result.Details = ResultDetails;
 Result.Button = ResultButton;
 Result.Tags = ResultTags;
+Result.Badge = Badge;
 
 export default Result;