+++ packages/doc/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.51.22](https://github.com/gympass/yoga/compare/@gympass/yoga-doc@2.51.21...@gympass/yoga-doc@2.51.22) (2024-05-27)
+
+**Note:** Version bump only for package @gympass/yoga-doc
+
+
+
+
+
+## [2.51.21](https://github.com/gympass/yoga/compare/@gympass/yoga-doc@2.51.20...@gympass/yoga-doc@2.51.21) (2024-05-27)
+
+**Note:** Version bump only for package @gympass/yoga-doc
+
+
+
+
+
+## [2.51.20](https://github.com/gympass/yoga/compare/@gympass/yoga-doc@2.51.19...@gympass/yoga-doc@2.51.20) (2024-05-23)
+
+**Note:** Version bump only for package @gympass/yoga-doc
+
+
+
+
+
+## [2.51.19](https://github.com/gympass/yoga/compare/@gympass/yoga-doc@2.51.18...@gympass/yoga-doc@2.51.19) (2024-05-22)
+
+**Note:** Version bump only for package @gympass/yoga-doc
+
+
+
+
+
+## [2.51.18](https://github.com/gympass/yoga/compare/@gympass/yoga-doc@2.51.17...@gympass/yoga-doc@2.51.18) (2024-05-22)
+
+**Note:** Version bump only for package @gympass/yoga-doc
+
+
+
+
+
 ## [2.51.17](https://github.com/gympass/yoga/compare/@gympass/yoga-doc@2.51.16...@gympass/yoga-doc@2.51.17) (2024-05-21)
 
 **Note:** Version bump only for package @gympass/yoga-doc