+++ packages/labnative/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.6.22](https://github.com/gympass/yoga/compare/@gympass/yoga-labnative@2.6.21...@gympass/yoga-labnative@2.6.22) (2024-05-27)
+
+**Note:** Version bump only for package @gympass/yoga-labnative
+
+
+
+
+
+## [2.6.21](https://github.com/gympass/yoga/compare/@gympass/yoga-labnative@2.6.20...@gympass/yoga-labnative@2.6.21) (2024-05-27)
+
+**Note:** Version bump only for package @gympass/yoga-labnative
+
+
+
+
+
+## [2.6.20](https://github.com/gympass/yoga/compare/@gympass/yoga-labnative@2.6.19...@gympass/yoga-labnative@2.6.20) (2024-05-23)
+
+**Note:** Version bump only for package @gympass/yoga-labnative
+
+
+
+
+
+## [2.6.19](https://github.com/gympass/yoga/compare/@gympass/yoga-labnative@2.6.18...@gympass/yoga-labnative@2.6.19) (2024-05-22)
+
+**Note:** Version bump only for package @gympass/yoga-labnative
+
+
+
+
+
+## [2.6.18](https://github.com/gympass/yoga/compare/@gympass/yoga-labnative@2.6.17...@gympass/yoga-labnative@2.6.18) (2024-05-22)
+
+**Note:** Version bump only for package @gympass/yoga-labnative
+
+
+
+
+
 ## [2.6.17](https://github.com/gympass/yoga/compare/@gympass/yoga-labnative@2.6.16...@gympass/yoga-labnative@2.6.17) (2024-05-21)
 
 **Note:** Version bump only for package @gympass/yoga-labnative